Add: 메인페이지 추합 메인페이지 컴포넌트 추합 수정 1
