+++ client/src/components/webItem/WebItem.styled.tsx
@@ -1,0 +1,12 @@
+import tw from 'tailwind-styled-components';
+
+export const WebItemContainer = tw.div`
+  w-80
+  relative
+  shadow-md
+  hover:bg-neutral-500/60
+`;
+
+export const WebItemImg = tw.img`
+  hover:opacity-80
+`;
