+++ client/src/pages/Main.tsx
@@ -1,7 +1,7 @@
 import Footer from '@/commons/Footer';
 import Header from '@/commons/atoms/header/Header';
 import CategoryNavBar from '@/commons/atoms/navbar/CategoryNavBar';
-import WebItem from '@/components/WebItem';
+import WebItem from '@/components/webItem/WebItem';
 
 export default function Main() {
   return (
